improve signature for `engine.matches`
A change in `adblock-rust` adds a field to signify that the matched
filter is `important`. This introduces a new boolean output argument to
`engine.matches` that represents whether or not a matched filter is
`important`.

Additionally, this changes `saved_from_exception` to just
`did_match_exception` and returns uses `true` as the return value only
when a filter matches. This way, it's possible to determine whether or
not an exception matched even if no filter was matched.